Add test to cover `last_login_at` timestamp in OAuth workflow (#3878)
* Add test to cover `last_login_at` timestamp in OAuth workflow

This is a follow-up to #3852. This PR above didn’t include a test for the expected behavior when logging in via OAuth. There have been a lot of changes to authentication testing on the develop branch that were not present in the 4.0 release branch, so any changes would have lead to annoying merge conflicts down the line.

But now that the 4.0 release is integrated, it’s time to add a proper test :)

* Fix OAuth test setup

This was an oversight and should have always called `super().setUp()` instead of `super().setUpClass()`. This didn’t fail when running tests for the entire file (`pytest test_sessions_api.py`), but it did always fail when running individual tests (`pytest test_sessions_api.py -k oauth_callback`).