Implement infinite scrolling for document search (#3855)
* Split up render method in PdfViewerSearch

This is just a small refactoring to make it easier to adjust this method.

* Rename fetchPage to fetchSearchResults

… because that is what it actually does -- it doesn’t load a page from the document, it loads a list of pages that contain the search terms.

* Implement infinite scrolling for document search results

Fixes #3320

* Remove useless code

I’m not sure what exactly the intention here was originally, but it didn’t have any effect in its current version. `page` and `res.index` would always be undefined and as a result, the `active` class was applied to every single link. The `active` class doesn’t apply any styles.

Probably this is just leftover from a previous iteration of the UI, so I’ve removed this.