Do not allow duplicate role memberships (#3815)
* Make name/email customizable in OAuth mocks

* Fix OAuth test setup

This was an oversight and should have always called `super().setUp()` instead of `super().setUpClass()`. This didn’t fail when running tests for the entire file (`pytest test_sessions_api.py`), but it did always fail when running individual tests (`pytest test_sessions_api.py -k oauth_callback`).

* Test OAuth group sync

* Log warning when trying to add duplicate role membership

This likely isn't the root cause for #3811, but it will help with debugging in the future.

* Add primary key constraint on `role_membership` table

While this doesn't fix the root cause for #3811, it ensures that this error doesn't result in inconsistent data that have previously permanently blocked users from sign-in. Having sensible constraints on tables is a good idea anyway.