Update Prometheus exporter probes so they do not actually generate all metrics (#3996)
* Update Prometheus exporter probes so they do not actually generate all metrics

Generating all metrics can take multiple seconds for large Aleph instances, which can make the liveness/readiness probe fail. The `name[]` query parameter can be used to select a subset of metrics the exporter should return (or no metrics as all in this case). That way, we can check that the exporter is up and running without actually generating any metrics.

* Slightly increase exporter default resources

This should be a little closer to real-world usage