Consider "layout: {}" to be empty
This shouldn't really change anything for Aleph as such, but Alfred
sends back "layout: {}". The reason for that is that it's rather a lot
of work to not send anything for these type of structs.

Go 1.24 will fix that next year, so I don't want to spend a lot of
effort on this. For now, this small change is needed to make it
compatible with the Alfred API.