House cleaning (#4021)
* Fix formatting

* Remove unused imports

* Remove unused assignments

* Fix linter ignore

This was ignored before already, but it didn’t apply because the comment wasn’t actually placed on the preceding line. It would be better to fix this properly, but that’s not possible without backend changes.

* Use strict type comparisons

* Fix typo (Ctl/Ctrl)

I think `Ctl` is occasionally used as well, but `Ctrl` is much more common and what you’ll find on most keyboards

* Use strict comparisons in table editor

* Use `forEach` instead of `map`

This caused a linter warning. `map` is only used here to iterate over the arrays, so `forEach` is the proper replacement.

* Remove unused files

These are leftovers from the time when react-ftm was a separate package

* Remove useless constructors

* Remove unused network diagram layout helper

The code in this file has been commented out for a long time, the file isn’t imported anywhere else, so I assume it’s safe to remove.

* Configure ESLint to fail if there are warnings

Now that all linter warnings are resolved, let’s make sure we don’t introduce any new ones.