Update network diagram entity viewer to avoid `componentWillReceiveProps`
The `EntityViewer` component displays a list of fields for the following entity properties:

1. Feature properties (no matter whether the property has a value or not)
2. Properties with at least one value
3. Properties manually added by the user using the "Add a property" select

The previous implementation kept a list of all visible properties (1-3) in the component state. It used the `componentWillReceiveProps` hook to reset this list when a user selected a different entity. The new implementation keeps only the selected properties (3) in the component state. Featured and non-empty properties (1+2) can be derived from the `entity` property at render time. In order to reset the list of selected properties when the user selects a different entity, we use the entity ID as a key. When the entity ID changes, React instantiates a new component (with fresh state) rather than reusing the component rendered for the previous entity.