bugfix: Point out use of Tags db and how it works
ingest-file uses the Tags functionality from servicelayer to cache certain document types
during ingest to speed up its operation. Normally the settings file defaults to trying
the ALEPH_DATABASE_URI followed by the FTM_STORE_URI, which in most cases should be set
to non-default values.

But if one were to just follow our prod environment docs blindly and not uncomment these
two settings then the TAGS_DATABASE_URI would be unset, causing servicelayer to pick the
default (sqlite:///) which breaks now that we have multiple threads writing to the tags
database.

This change documents how this setting works and encourages one to set it if the above
two settings are left to defaults.

Relates to #4002, #3816